requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.23.10
google-generativeai>=0.4.0
chromadb>=0.4.0
python-dotenv>=1.0.0
numpy>=1.24.0
opencv-python>=4.8.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.0.0
